Crate lyon

Lyon

2d rendering on the GPU in rust.

Crates

lyon_tessellation (documentation): Path tessellation routines.

lyon_path_builder (documentation): Tools to facilitate building paths.

lyon_path_iterator (documentation): Tools to facilitate iteratring over paths.

lyon_path (documentation): A simple optional path datat structure, provided for convenience.

lyon_bezier (documentation): Cubic and Quadratic 2d bezier math.

lyon_extra (documentation): Additional testing and debugging tools.

lyon_core (documentation): Common types to most lyon crates.

This crate is just a meta-crate, reexporting the crates listed above.

Additional documentation and links

There is some useful documentaion on the project's wiki.

The source code is available on the project's git repository.

Interested in contributing? This page is is probably what you are looking for! You can also look at the list of issues marked as easy, they are a good place to start.

Reexports

pub extern crate lyon_core;

pub extern crate lyon_path;

pub extern crate lyon_path_builder;

pub extern crate lyon_path_iterator;

pub extern crate lyon_tessellation;

pub extern crate lyon_bezier;

pub extern crate lyon_extra;

pub extern crate lyon_svg;

pub use lyon_core::*;

Modules

bezier
Bezier curve related maths and tools.

extra

path

path_builder

path_iterator

svg

tessellation