Module lyon::bezier

Structs

CubicBezierSegment
A 2d curve segment defined by four points: the beginning of the segment, two control points and the end of the segment.

CubicFlatteningIter
An iterator over a cubic bezier segment that yields line segments approximating the curve for a given approximation threshold.

QuadraticBezierSegment
A 2d curve segment defined by three points: the beginning of the segment, a control point and the end of the segment.

QuadraticFlatteningIter
An iterator over a quadratic bezier segment that yields line segments approximating the curve for a given approximation threshold.

Type Definitions

Point

Vec2