Module lyon::extra::triangle_rasterizer

Structs

ColorTarget

Traits

GetColor
Implemented vertices and constants that can return a color.

PixelShader
An operation that is applied to each rasterized pixel

ShadingStage

VertexData

Functions

rasterize_triangles
A software triangle rasterizer intended for ref testing and to help debugging the output of the various tessellation routines.