Module lyon::path_builder

Structs

FlatteningBuilder
Generates flattened paths

SvgPathBuilder
Implements the Svg building interface on top of a PathBuilder.

Traits

BaseBuilder
The most basic path building interface. Does not handle any kind of curve.

PathBuilder
The main path building interface. More elaborate interfaces are built on top of the provided primitives.

PolygonBuilder
Build a path from a simple list of points.

SvgBuilder
A path building interface that tries to stay close to SVG's path specification. https://svgwg.org/specs/paths/