Module lyon::path_iterator

Lyon path builder

Tools to iterate over path objects.

Overview

TODO

Examples

TODO

Structs

FlatteningIter
An iterator that consumes an PathIterator and yields FlattenedEvents.

PathStateIter
An adapater iterator that implements PathIterator on top of an Iterator.

PathStateSvgIter
An adapater iterator that implements SvgIterator on top of an Iterator.

SegmentIterator
Consumes an iterator of path events and yields segments.

SvgToPathIter

Enums

Segment
Convenience for algorithms which prefer to iterate over segments directly rather than path events.

Traits

FlattenedIterator
An extension to the common Iterator interface, that adds information which is useful when chaining path-specific iterators.

PathIterator
An extension to the common Iterator interface, that adds information which is useful when chaining path-specific iterators.

SvgIterator
An extension to the common Iterator interface, that adds information which is useful when chaining path-specific iterators.