Module lyon::tessellation

Modules

basic_shapes
Tessellation routines for simple shapes.

events

fixed

geometry_builder

math

math_utils
Various math tools that are usefull for several modules.

path_fill

path_state

path_stroke

Structs

ArcFlags
Flag parameters for arcs as described by the SVG specification.

PathState
Represents the current state of a path while it is being built.

Enums

FlattenedEvent

PathEvent

SvgEvent