Module lyon::tessellation::fixed

Structs

Fp32
A 32 fixed point number. The size of the fractional is defined by the type parameter F.

Fp64
A 64 bits fixed point number. The size of the fractional is defined by the type parameter F.

_1

_10

_11

_12

_13

_14

_15

_16

_2

_24

_3

_32

_4

_5

_6

_7

_8

_9

Traits

FractionalBits