Struct lyon::tessellation::math::TypedPoint2D

pub struct TypedPoint2D<T, U> {
    pub x: T,
    pub y: T,
    // some fields omitted
}

A 2d Point tagged with a unit.

Fields
x: T y: T
Methods

impl<T, U> TypedPoint2D<T, U> where T: Copy + Zero

fn zero() -> TypedPoint2D<T, U>

Constructor, setting all components to zero.

fn to_3d(&self) -> TypedPoint3D<T, U>

Convert into a 3d point.

impl<T, U> TypedPoint2D<T, U> where T: Copy

fn new(x: T, y: T) -> TypedPoint2D<T, U>

Constructor taking scalar values directly.

fn from_lengths(x: Length<T, U>, y: Length<T, U>) -> TypedPoint2D<T, U>

Constructor taking properly typed Lengths instead of scalar values.

fn x_typed(&self) -> Length<T, U>

Returns self.x as a Length carrying the unit.

fn y_typed(&self) -> Length<T, U>

Returns self.y as a Length carrying the unit.

fn to_untyped(&self) -> TypedPoint2D<T, UnknownUnit>

Drop the units, preserving only the numeric value.

fn from_untyped(p: &TypedPoint2D<T, UnknownUnit>) -> TypedPoint2D<T, U>

Tag a unitless value with units.

fn to_array(&self) -> [T; 2]

impl<T, U> TypedPoint2D<T, U> where T: Mul<T, Output=T> + Add<T, Output=T> + Sub<T, Output=T> + Copy

fn dot(self, other: TypedPoint2D<T, U>) -> T

Dot product.

fn cross(self, other: TypedPoint2D<T, U>) -> T

Returns the norm of the cross product [self.x, self.y, 0] x [other.x, other.y, 0]..

impl<T, U> TypedPoint2D<T, U> where T: Copy + Add<T, Output=T>

fn add_size(&self, other: &TypedSize2D<T, U>) -> TypedPoint2D<T, U>

impl<T, U> TypedPoint2D<T, U> where T: Float

fn min(self, other: TypedPoint2D<T, U>) -> TypedPoint2D<T, U>

fn max(self, other: TypedPoint2D<T, U>) -> TypedPoint2D<T, U>

impl<T, U> TypedPoint2D<T, U> where T: Round

fn round(&self) -> TypedPoint2D<T, U>

Rounds each component to the nearest integer value.

This behavior is preserved for negative values (unlike the basic cast). For example { -0.1, -0.8 }.round() == { 0.0, -1.0 }

impl<T, U> TypedPoint2D<T, U> where T: Ceil

fn ceil(&self) -> TypedPoint2D<T, U>

Rounds each component to the smallest integer equal or greater than the orginal value.

This behavior is preserved for negative values (unlike the basic cast). For example { -0.1, -0.8 }.ceil() == { 0.0, 0.0 }.

impl<T, U> TypedPoint2D<T, U> where T: Floor

fn floor(&self) -> TypedPoint2D<T, U>

Rounds each component to the biggest integer equal or lower than the orginal value.

This behavior is preserved for negative values (unlike the basic cast). For example { -0.1, -0.8 }.floor() == { -1.0, -1.0 }.

impl<T, U> TypedPoint2D<T, U> where T: Copy + NumCast

fn cast<NewT>(&self) -> Option<TypedPoint2D<NewT, U>> where NewT: Copy + NumCast

Cast from one numeric representation to another, preserving the units.

When casting from floating point to integer coordinates, the decimals are truncated as one would expect from a simple cast, but this behavior does not always make sense geometrically. Consider using round(), ceil or floor() before casting.

fn to_f32(&self) -> TypedPoint2D<f32, U>

Cast into an f32 vector.

fn to_uint(&self) -> TypedPoint2D<usize, U>

Cast into an usize point, truncating decimals if any.

When casting from floating point vectors, it is worth considering whether to round(), ceil() or floor() before the cast in order to obtain the desired conversion behavior.

fn to_i32(&self) -> TypedPoint2D<i32, U>

Cast into an i32 point, truncating decimals if any.

When casting from floating point vectors, it is worth considering whether to round(), ceil() or floor() before the cast in order to obtain the desired conversion behavior.

fn to_i64(&self) -> TypedPoint2D<i64, U>

Cast into an i64 point, truncating decimals if any.

When casting from floating point vectors, it is worth considering whether to round(), ceil() or floor() before the cast in order to obtain the desired conversion behavior.