Struct lyon::tessellation::path_fill::FillOptions

pub struct FillOptions {
    pub tolerance: f32,
    pub fill_rule: FillRule,
    pub vertex_aa: bool,
    // some fields omitted
}

Parameters for the tessellator.

Fields
tolerance: f32

Maximum allowed distance to the path when building an approximation.

fill_rule: FillRule

See the SVG specification.

Currently, only the EvenOdd rule is implemented.

vertex_aa: bool

An anti-aliasing trick extruding a 1-px wide strip around the edges with a gradient to smooth the edges.

Not implemented yet!

Methods

impl FillOptions

fn default() -> FillOptions

fn even_odd() -> FillOptions

fn non_zero() -> FillOptions

fn with_tolerance(self, tolerance: f32) -> FillOptions

fn with_vertex_aa(self) -> FillOptions