Struct lyon::tessellation::path_fill::FillTessellator

pub struct FillTessellator {
    // some fields omitted
}

A Context object that can tessellate fill operations for complex paths.

The Tessellator API is not stable yet. For example it is not clear whether we will use separate Tessellator structs for some of the different configurations (vertex-aa, etc), or if evertything can be implemented with the same algorithm.

Methods

impl FillTessellator

fn new() -> FillTessellator

Constructor.

fn tessellate_events<Output>(&mut self, events: &FillEvents, options: &FillOptions, output: &mut Output) -> Result<Count, FillError> where Output: BezierGeometryBuilder<TypedPoint2D<f32, UnknownUnit>>

Compute the tessellation.

fn enable_logging(&mut self)

Enable some verbose logging during the tessellation, for debugging purposes.